requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
plotly>=5.18
streamlit>=1.30
reportlab>=4.0
pywebview>=5.0; platform_system=="Windows"
